test: 移除 ExportPdfTest 基类
